Share single Spring context across the configprops integration tests
