+++ sbs/axelix-spring-boot-2-starter/src/test/java/com/axelixlabs/axelix/sbs/spring/core/configprops/DefaultConfigurationPropertiesServiceTest.java
@@ -17,36 +17,23 @@
  */
 package com.axelixlabs.axelix.sbs.spring.core.configprops;
 
-import java.util.List;
-import java.util.Map;
-import java.util.Set;
-import java.util.stream.Collectors;
-
-import org.jspecify.annotations.Nullable;
-import org.junit.jupiter.api.Nested;
-import org.junit.jupiter.api.Test;
-
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.context.properties.ConfigurationProperties;
-import org.springframework.boot.context.properties.EnableConfigurationProperties;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.boot.test.context.TestConfiguration;
-import org.springframework.context.ApplicationContext;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Import;
-
 import com.axelixlabs.axelix.common.api.ConfigurationPropertiesFeed;
 import com.axelixlabs.axelix.common.api.KeyValue;
 import com.axelixlabs.axelix.common.auth.core.DefaultAuthority;
 import com.axelixlabs.axelix.common.auth.core.DefaultSecurityContext;
 import com.axelixlabs.axelix.common.auth.core.SecurityContext;
-import com.axelixlabs.axelix.common.auth.core.SecurityContextExecutor;
 import com.axelixlabs.axelix.common.auth.core.User;
-import com.axelixlabs.axelix.sbs.spring.core.auth.RequiredAuthorityCheckService;
 import com.axelixlabs.axelix.sbs.spring.core.auth.ThreadLocalSecurityContextExecutor;
-import com.axelixlabs.axelix.sbs.spring.core.config.EndpointsConfigurationProperties;
-import com.axelixlabs.axelix.sbs.spring.core.env.DefaultPropertyNameNormalizer;
-import com.axelixlabs.axelix.sbs.spring.core.env.PropertyNameNormalizer;
+import org.jspecify.annotations.Nullable;
+import org.junit.jupiter.api.Nested;
+import org.junit.jupiter.api.Test;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.beans.factory.annotation.Qualifier;
+
+import java.util.List;
+import java.util.Map;
+import java.util.Set;
+import java.util.stream.Collectors;
 
 import static com.axelixlabs.axelix.sbs.spring.core.utils.UserUtils.createUserWithAuthorities;
 import static org.assertj.core.api.Assertions.assertThat;
@@ -63,12 +50,11 @@ public class DefaultConfigurationPropertiesServiceTest {
 
     private final ThreadLocalSecurityContextExecutor securityContextExecutor = new ThreadLocalSecurityContextExecutor();
 
-    @SpringBootTest
     @Nested
-    @Import(TestConfigWithAllPropertiesSanitized.class)
-    class WithoutExplicitSanitizationProperties {
+    class WithoutExplicitSanitizationProperties extends AbstractConfigurationPropertiesTest {
 
         @Autowired
+        @Qualifier("configurationPropertiesServiceAll")
         private ConfigurationPropertiesService configurationPropertiesService;
 
         @Test
@@ -110,17 +96,11 @@ public class DefaultConfigurationPropertiesServiceTest {
         }
     }
 
-    @SpringBootTest(
-            properties = {
-                "axelix.prop.test.tags.forSanitization=toBeSanitized",
-                "axelix.prop.test.tags.FOR_SANITIZATION=toBeSanitized"
-            })
     @Nested
-    @EnableConfigurationProperties(AxelixConfigurationProperties.class)
-    @Import(TestConfigWithExplicitSanitizationProperties.class)
-    class WithExplicitSanitizationProperties {
+    class WithExplicitSanitizationProperties extends AbstractConfigurationPropertiesTest {
 
         @Autowired
+        @Qualifier("configurationPropertiesServiceExplicit")
         private ConfigurationPropertiesService configurationPropertiesService;
 
         @Test
@@ -168,91 +148,6 @@ public class DefaultConfigurationPropertiesServiceTest {
 
             assertThat(values).doesNotContain("******");
             assertThat(configProps).isNotNull().isInstanceOf(ConfigurationPropertiesFeed.class);
-        }
-    }
-
-    @TestConfiguration
-    @Import(ConfigurationPropertiesServiceTestConfiguration.class)
-    static class TestConfigWithAllPropertiesSanitized {
-
-        @Bean
-        public SmartSanitizingFunction smartSanitizingFunction(PropertyNameNormalizer propertyNameNormalizer) {
-            return new SmartSanitizingFunction(EndpointsConfigurationProperties.SANITIZE_ALL, propertyNameNormalizer);
-        }
-    }
-
-    @TestConfiguration
-    @Import(ConfigurationPropertiesServiceTestConfiguration.class)
-    static class TestConfigWithExplicitSanitizationProperties {
-
-        @Bean
-        public SmartSanitizingFunction smartSanitizingFunction(PropertyNameNormalizer propertyNameNormalizer) {
-            return new SmartSanitizingFunction(
-                    List.of("axelix.prop.test.tags.forSanitization", "axelix.prop.test.tags.FOR_SANITIZATION"),
-                    propertyNameNormalizer);
-        }
-    }
-
-    @ConfigurationProperties(prefix = "axelix.prop.test")
-    static class AxelixConfigurationProperties {
-
-        private Map<String, String> tags;
-
-        public Map<String, String> getTags() {
-            return tags;
-        }
-
-        public void setTags(Map<String, String> tags) {
-            this.tags = tags;
-        }
-    }
-
-    static class ConfigurationPropertiesServiceTestConfiguration {
-
-        @Bean
-        @ConfigurationProperties(prefix = "axelix.sbs.endpoints.config")
-        public EndpointsConfigurationProperties endpointsConfigurationProperties() {
-            return new EndpointsConfigurationProperties();
-        }
-
-        @Bean
-        public ConfigurationPropertiesFlattener configurationPropertiesFlattener() {
-            return new DefaultConfigurationPropertiesFlattener();
-        }
-
-        @Bean
-        public ConfigurationPropertiesConverter configurationPropertiesConverter(
-                ConfigurationPropertiesFlattener configurationPropertiesFlattener) {
-            return new DefaultConfigurationPropertiesConverter(configurationPropertiesFlattener);
-        }
-
-        @Bean
-        public PropertyNameNormalizer propertyNameNormalizer() {
-            return new DefaultPropertyNameNormalizer();
-        }
-
-        @Bean
-        public SecurityContextExecutor securityContextExecutor() {
-            return new ThreadLocalSecurityContextExecutor();
-        }
-
-        @Bean
-        public RequiredAuthorityCheckService requiredAuthorityCheckService(
-                SecurityContextExecutor securityContextExecutor) {
-            return new RequiredAuthorityCheckService(securityContextExecutor);
-        }
-
-        @Bean
-        public ConfigurationPropertiesService configurationPropertiesService(
-                SmartSanitizingFunction smartSanitizingFunction,
-                ApplicationContext applicationContext,
-                ConfigurationPropertiesConverter configurationPropertiesConverter,
-                RequiredAuthorityCheckService requiredAuthorityCheckService) {
-            return new DefaultConfigurationPropertiesService(
-                    smartSanitizingFunction,
-                    applicationContext,
-                    configurationPropertiesConverter,
-                    requiredAuthorityCheckService);
         }
     }
 }
